# Apply hard limit to partition range vectors in secondary index queries