@ndragazis 2025-02-05
v2 Changelog:
* Adjust the the unit test to succeed both before and after the fix.
* Disable the unit test in "sanitize" mode (ASAN messes up with memory allocations).
* Increase threshold in unit test to 512KiB to filter out large allocations that may be caused by other internal operations.
* Wrap `paging_adjusted_options` with `do_with()` (fixes use-after-free bug).
* Hardcode the max vector size to 1000 instead of deriving it from the allocator's preferred max.
* Add runtime check with `on_fatal_internal_error()` to ensure that 1000 won't cause large allocations.