@ndragazis 2025-02-07
## Measurements

@avikivity, @denesb, I measured the latency of some secondary index queries on a local 3-node cluster.

Setup:
* cluster: 3 nodes, 1 vcpu, 4G memory, dev mode
* schema: replication factor = 3, no tablets (secondary indexed not supported yet)
* query: indexed, selects all rows, QL = LOCAL_QUORUM
* 5 runs per measurement

Requirements:
* [simple_scylla_cluster_manager.py](https://github.com/scylladb/scylla-dev-utils/blob/master/simple_scylla_cluster_manager.py) to set up a local cluster
* custom [test scripts](https://github.com/ndragazis/scylladb/tree/perf-secondary-index) to create the table, fill it with data, run queries that retrieve the whole table with various page sizes, and print latency stats
* scylla braches for testing:
   - master (`2025.2.0~dev-0.20250207.2a9949475285`)
   - current PR: limited vector size of 1000 partition keys
   - previous PR with internal paging (https://github.com/scylladb/scylladb/pull/22254): internal page size of 1000

Steps:
1. Build scylla from desired branch.
2. Spin up the cluster:
    ```bash
    cd scylla-dev-utils/
    mkdir rundir && cd rundir
    ../simple_scylla_cluster_manager.py create-cluster -n3 -c1 -s -p /home/ndragazis/Documents/scylladb/git/scylladb
    ```
3. Find the IP of one of the nodes:
    ```bash
    ./simple_scylla_cluster_manager.py status | jq -r .[0].ip
    ```
4. Edit the IP in the test script.
5. Run the test script.

## Results

### Case 1: Small partition range vectors, no short reads

1,000 partitions, 100 rows per partition
<details>
<summary>Schema:</summary>

```sql
CREATE KEYSPACE ks WITH replication = {'class': 'org.apache.cassandra.locator.NetworkTopologyStrategy', 'datacenter1': '3'} AND durable_writes = true AND tablets = {'enabled': false};

CREATE TABLE ks.t1 (
    pk1 int,
    pk2 int,
    ck int,
    value int,
    PRIMARY KEY ((pk1, pk2), ck)
) WITH CLUSTERING ORDER BY (ck ASC)  
    AND bloom_filter_fp_chance = 0.01
    AND caching = {'keys': 'ALL', 'rows_per_partition': 'ALL'}
    AND comment = ''
    AND compaction = {'class': 'IncrementalCompactionStrategy'}
    AND compression = {'sstable_compression': 'org.apache.cassandra.io.compress.LZ4Compressor'}
    AND crc_check_chance = 1
    AND default_time_to_live = 0
    AND gc_grace_seconds = 864000    
    AND max_index_interval = 2048      
    AND memtable_flush_period_in_ms = 0
    AND min_index_interval = 128
    AND speculative_retry = '99.0PERCENTILE'
   AND tombstone_gc = {'mode': 'timeout', 'propagation_delay_in_seconds': '3600'};

CREATE INDEX t1_pk2_idx ON ks.t1(pk2);
```

</details>

```
+------------+-------------------+-------------------+-------------------+
|  Page Size |      Master       |   Vector Limit    |  Internal Paging  |
+============+===================+===================+===================+
|            |   Latency (sec)   |   Latency (sec)   |   Latency (sec)   |
+------------+-------------------+-------------------+-------------------+
|     100    |  4.06 ± 0.13      |  4.12 ± 0.16      |  4.15 ± 0.13      |
+------------+-------------------+-------------------+-------------------+
|    1000    |  2.90 ± 0.08      |  2.94 ± 0.06      |  2.94 ± 0.07      |
+------------+-------------------+-------------------+-------------------+
|    2000    |  2.72 ± 0.06      |  2.77 ± 0.05      |  2.76 ± 0.05      |
+------------+-------------------+-------------------+-------------------+
|    5000    |  2.63 ± 0.08      |  2.61 ± 0.08      |  2.67 ± 0.03      |
+------------+-------------------+-------------------+-------------------+
|   10000    |  2.54 ± 0.07      |  2.57 ± 0.10      |  2.63 ± 0.05      |
+------------+-------------------+-------------------+-------------------+
|     -1     |  2.55 ± 0.03      |  2.50 ± 0.04      |  2.55 ± 0.08      |
+------------+-------------------+-------------------+-------------------+
```
Latencies are pretty much the same in all cases. The vector limit is expected to perform a bit faster than internal paging, because it fetches all primary keys from the index view in one request instead of `<page size> / 1000` requests. This is hardly noticeable in the measurements.

### Case 2: Large partition range vectors, no short reads

30,000 partitions, 1 row per partition
Same schema as case 1.

```
+------------+-------------------+-------------------+-------------------+
|  Page Size |      Master       |   Vector Limit    |  Internal Paging  |
+============+===================+===================+===================+
|            |   Latency (sec)   |   Latency (sec)   |   Latency (sec)   |
+------------+-------------------+-------------------+-------------------+
|     100    |  5.80 ± 0.13      |  5.64 ± 0.10      |  5.81 ± 0.05      |
+------------+-------------------+-------------------+-------------------+
|    1000    |  4.77 ± 0.07      |  4.62 ± 0.06      |  4.84 ± 0.09      |
+------------+-------------------+-------------------+-------------------+
|    2000    |  4.67 ± 0.07      |  5.13 ± 0.03      |  4.76 ± 0.04      |
+------------+-------------------+-------------------+-------------------+
|    5000    |  4.82 ± 0.09      |  6.25 ± 0.06      |  4.86 ± 0.16      |
+------------+-------------------+-------------------+-------------------+
|   10000    |  4.89 ± 0.36      |  7.52 ± 0.13      |  5.09 ± 0.30      |
+------------+-------------------+-------------------+-------------------+
|     -1     |  4.90 ± 0.67      |  4.79 ± 0.33      |  4.92 ± 0.47      |
+------------+-------------------+-------------------+-------------------+
```
With the vector limit, the querier cache is lost above 1000, so latencies increase. This doesn't happen with internal paging.

### Case 3: Small partition range vectors, short base table reads

1,000 partitions, 100 rows per partition, ~1KiB per row (large non-primary key column)
Same schema as case 1, except that the non-primary key column is a blob with 1KiB of random data.

```
+------------+-------------------+-------------------+-------------------+
|  Page Size |      Master       |   Vector Limit    |  Internal Paging  |
+============+===================+===================+===================+
|            |   Latency (sec)   |   Latency (sec)   |   Latency (sec)   |
+------------+-------------------+-------------------+-------------------+
|     100    |  4.35 ± 0.09      |  4.45 ± 0.05      |  4.51 ± 0.11      |
+------------+-------------------+-------------------+-------------------+
|    1000    |  3.20 ± 0.08      |  3.13 ± 0.11      |  3.22 ± 0.06      |
+------------+-------------------+-------------------+-------------------+
|    2000    |  3.00 ± 0.11      |  2.98 ± 0.08      |  3.07 ± 0.06      |
+------------+-------------------+-------------------+-------------------+
|    5000    |  3.77 ± 0.16      |  3.74 ± 0.09      |  3.36 ± 0.03      |
+------------+-------------------+-------------------+-------------------+
|   10000    |  5.79 ± 0.16      |  5.86 ± 0.21      |  3.36 ± 0.02      |
+------------+-------------------+-------------------+-------------------+
|     -1     |  2.99 ± 0.12      |  2.93 ± 0.12      |  2.90 ± 0.07      |
+------------+-------------------+-------------------+-------------------+
```
In this test, each partition is 100KiB and contains 100 rows, so given the exponential concurrency on base table reads, we expect to hit a short read when a read exceeds 1MiB, so: (1 * 100KiB) + (2 * 100KiB) + (4 * 100KiB) + (8 * 100KiB) + **(16 * 100KiB)** = 3100 KiB -> 3100 rows -> 5000 page size. Indeed, internal paging seems to perform better than the vector limit above 2000. That's because, with internal paging, the number of discarded partition keys are bounded (in this case 1000 - 3100 % 1000 = 900). In contrast, with the vector limit, we discard 5000 - 3100 = 1900 and 10000 - 3100 = 6900 partition keys. Notice also that internal paging has a slight overhead above 2000. This must be due to losing the querier cache.

PS: I don't understand why the coordinator declares a read as short only based on the last read, and not from the aggregate of all past reads of this query. In the above example, the short read happens way beyond 1MiB.